### 2017-08-30 #386 @whedon
[PRE REVIEW]: SimuPy: A python framework for modeling and simulating dynamical systems
**Submitting author:** @sixpearls (<a href="http://orcid.org/0000-0001-5602-1888">Benjamin Margolis</a>)
**Repository:** <a href="https://github.com/simupy/simupy" target ="_blank">https://github.com/simupy/simupy</a>
**Version:** v1.0.0
