### 2017-08-30 #386 @sixpearls
In terms of preferred languages and subject area, @moorepants is probably the best suited reviewer. However, we're at the same institution (and we know each other); I don't know what the JOSS policy is for this situation. In general I would trust @moorepants's ability to determine if he could be fair, especially so with the open discussion format. There are also a number of other reviewers who should be able to provide good feedback in addition to @moorepants.

If possible, I would appreciate it if the reviewer would be willing to work with me to get a DOI by Sept 8th so I can reference it in a conference manuscript. That's pretty soon, so I don't expect that to be treated as a hard constraint.

Thanks!
Ben
