+++ docs/index.rst
@@ -3,6 +3,9 @@ SimuPy
 
 A framework for modeling and simulating dynamical systems. This provides an open source, python-based alternative to Simulink.
 
+.. image:: docs/_static/simupy.png
+   :alt: SimuPy Logo
+
 .. toctree::
    :maxdepth: 2
 
