+++ docs/index.rst
@@ -3,8 +3,9 @@ SimuPy
 
 A framework for modeling and simulating dynamical systems. This provides an open source, python-based alternative to Simulink.
 
-.. image:: docs/_static/simupy.png
+.. image:: docs/_static/simupy_color.png
    :alt: SimuPy Logo
+   :scale: 75%
 
 .. toctree::
    :maxdepth: 2
