### 2017-09-01 2ebf457
remove bar
