+++ README.rst
@@ -4,8 +4,6 @@
 .. |simupy_logo| image:: docs/_static/simupy_color_small.png
    :alt: SimuPy
 
-|
-
 .. image:: https://img.shields.io/pypi/v/simupy.svg
     :alt: PyPI Package latest release
     :target: https://pypi.python.org/pypi/simupy
