### 2017-09-01 1fa53d9
docs updates
