+++ docs/index.rst
@@ -1,15 +1,11 @@
 SimuPy
 ======
 
-A framework for modeling and simulating dynamical systems. This provides an open source, python-based alternative to Simulink.
-
-.. image:: docs/_static/simupy_color.png
-   :alt: SimuPy Logo
-   :scale: 75%
+A Python framework for modeling and simulating dynamical systems.
 
 .. toctree::
    :maxdepth: 2
 
-   overview
+   SimuPy <overview>
    math
    api/api
