### 2017-09-01 f46e1d8
fix logo to use in docs
