### 2017-09-05 #386 @cMadan
@sixpearls, I'm happy to serve as editor for this submission :)

@moorepants, if you think you can objectively review this project (and are otherwise available), I think it's fine to have you as a reviewer here. If not, perhaps @sixpearls can suggest someone else from the list linked in the first comment here.
