### 2017-09-05 #386 @sixpearls
Thanks, @moorepants. I understand the preference!

It looks like @lheagy, @kyleniemeyer, and @katyhuff are all python users who might be familiar with this type of simulation. Perhaps one of you might be willing to review this?
