### 2017-09-06 #386 @kyleniemeyer
@cMadan @sixpearls I could also probably review this, but I'd like to recommend finding a reviewer that isn't a JOSS editor... our plates are already pretty full.
