### 2017-09-06 #386 @sixpearls
@katyhuff @kyleniemeyer @lheagy thanks for your responses. Appreciate you all taking the time to respond. Thanks for the suggestion about cross-referencing the editorial board list.

From the list, it looks like @arghdos and @mosteo work in related areas and use python. Beyond that, it would start to be at least a little out of field and/or preferred programming language based on the suggested reviewer list. I'm not sure how much that matters.

@moorepants thanks for the suggestion!

@demotu @siboles @stavness @melund @jslee02 @cybanical @chrisdembia would any of you be willing to review [my submission](https://github.com/simupy/simupy) to [JOSS](http://joss.theoj.org/)? I believe you're all familiar with the journal, but the review guidelines are [here](http://joss.theoj.org/about#reviewer_guidelines).
