### 2017-09-06 f537f2e
add aliases for LTI system matrices
