### 2017-09-06 b35b76e
remove descriptor, linting
