+++ docs/api/descriptor.rst
@@ -1,7 +1,0 @@
-descriptor module
-=================
-
-.. automodule:: simupy.descriptor
-    :members:
-    :undoc-members:
-    :show-inheritance:
