+++ docs/conf.py
@@ -19,9 +19,9 @@
 #
 import os
 import sys
-import simupy
 
 sys.path.insert(0, os.path.abspath('..'))
+import simupy
 
 # -- General configuration ------------------------------------------------
 
