### 2017-09-06 #386 @demotu
Yes, I can review your submission; the simupy sounds very interesting.

@demotu <https://github.com/demotu>

-
Marcos Duarte
http://demotu.org/

On Wed, Sep 6, 2017 at 1:26 AM, Ben Margolis <notifications@github.com>
wrote:

> @katyhuff <https://github.com/katyhuff> @kyleniemeyer
> <https://github.com/kyleniemeyer> @lheagy <https://github.com/lheagy>
> thanks for your responses. Appreciate you all taking the time to respond.
> Thanks for the suggestion about cross-referencing the editorial board list.
>
> From the list, it looks like @arghdos <https://github.com/arghdos> and
> @mosteo <https://github.com/mosteo> work in related areas and use python.
> Beyond that, it would start to be at least a little out of field and/or
> preferred programming language based on the suggested reviewer list. I'm
> not sure how much that matters.
>
> @moorepants <https://github.com/moorepants> thanks for the suggestion!
>
> @demotu <https://github.com/demotu> @siboles <https://github.com/siboles>
> @stavness <https://github.com/stavness> @melund
> <https://github.com/melund> @jslee02 <https://github.com/jslee02>
> @cybanical <https://github.com/cybanical> @chrisdembia
> <https://github.com/chrisdembia> would any of you be willing to review my
> submission <https://github.com/simupy/simupy> to JOSS
> <http://joss.theoj.org/>? I believe you're all familiar with the journal,
> but the review guidelines are here
> <http://joss.theoj.org/about#reviewer_guidelines>.
>
