### 2017-09-10 cc9cdd7
use plt.show instead of interactive plotting for examples
