+++ examples/simple_saturation.py
@@ -36,3 +36,4 @@ plt.xlabel('$t$, s')
 plt.ylabel('$x(t)$')
 plt.title('simple saturation demonstration')
 plt.tight_layout()
+plt.show()
