+++ examples/vanderpol.py
@@ -35,6 +35,7 @@ plt.ylabel('$x_i(t)$')
 plt.xlabel('$t$, s')
 plt.title('system state vs time')
 plt.tight_layout()
+plt.show()
 
 plt.figure()
 plt.plot(*res.x.T)
@@ -42,6 +43,7 @@ plt.xlabel('$x_1(t)$')
 plt.ylabel('$x_2(t)$')
 plt.title('phase portrait of system')
 plt.tight_layout()
+plt.show()
 
 plt.figure()
 plt.plot(res.t, res.y)
@@ -49,3 +51,4 @@ plt.legend([r'$\left| \mathbf{x}(t) \right|$', r'$\angle \mathbf{x} (t)$'])
 plt.xlabel('$t$, s')
 plt.title('system outputs vs time')
 plt.tight_layout()
+plt.show()
