### 2017-09-10 10d286e
disconnected systems should be connected to zeros
