### 2017-09-10 9d694e4
set to version 0.2.0 for pre release
