+++ simupy/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.1.0.dev1'
+__version__ = '0.2.0'
