### 2017-09-10 49d227b
fix manifest
