+++ setup.py
@@ -13,7 +13,6 @@ with open(path.join(here, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
 
 long_description = re.sub(r':doc:`(([^<]+))[^`]+`', r'\2', long_description)
-# long_description = "\nSimuPy"+long_description.split("|simupy_logo|")[-1]
 
 setup(
     name='simupy',
