### 2017-09-11 #396 @demotu
The submission addresses all JOSS points sufficiently.
I installed simupy and ran the examples. Everything works fine.
The link for the API doc at the README.rst doesn't work.
Could you provide working links for the documentation and docstrings?
