### 2017-09-12 #396 @sixpearls
Hi @demotu, thanks for your feedback, I'm glad you like the software!

The link to the API docs on the README is intentionally broken because the README is imported into a page of the [real docs](https://simupy.readthedocs.io/en/latest/index.html) and I wanted to keep the documentation DRY (Do not Repeat Yourself). Github does render a link to the documentation on readthedocs as a badge correctly at the top of the file.

Do you think this is sufficient?
