### 2017-09-12 #396 @demotu
Using Sympy as example, it has a link to its homepage at the very first
line and a specific session about documentation and usage with a clear link
to the docs. Particularly about documentation, I wouldn't mind redundancy.

-
Marcos Duarte
http://demotu.org/

On Mon, Sep 11, 2017 at 9:17 PM, Ben Margolis <notifications@github.com>
wrote:

> Hi @demotu <https://github.com/demotu>, thanks for your feedback, I'm
> glad you like the software!
>
> The link to the API docs on the README is intentionally broken because the
> README is imported into a page of the real docs
> <https://simupy.readthedocs.io/en/latest/index.html> and I wanted to keep
> the documentation DRY (Do not Repeat Yourself). Github does render a link
> to the documentation on readthedocs as a badge correctly at the top of the
> file.
>
> Do you think this is sufficient?
>
