### 2017-09-13 7fe396c
provide README link to API docs for github and pypi
