+++ docs/overview.rst
@@ -1,1 +1,3 @@
 .. include:: ../README.rst
+
+.. |API documentation| replace:: :doc:`API Documentation<api/api>`
