### 2017-09-13 e522246
make links better on pypi
