### 2017-09-13 #396 @sixpearls
Thanks, I agree working links to the documentation everywhere is helpful.

I figured out how to do it in a way I'm happy with. Now, the github README will point to the latest docs on readthedocs, the pypi description will point to the most recent X.Y.Z build of the docs on readthedocs, and a built version of the docs will have a local link.

Does this address all your concerns?
