### 2017-09-13 #396 @cMadan
@sixpearls, I looked the repo over and everything looks great! You did an excellent job of preparing a useful readme, paper.md, and examples.

@demotu, is there a reason you haven't checked the performance checkbox in the list in the first comment here?
