### 2017-09-13 #396 @demotu
I just checked the performance checkbox.

-
Marcos Duarte
http://demotu.org/

On Wed, Sep 13, 2017 at 7:59 AM, Christopher Madan <notifications@github.com
> wrote:

> @sixpearls <https://github.com/sixpearls>, I looked the repo over and
> everything looks great! You did an excellent job of preparing a useful
> readme, paper.md, and examples.
>
> @demotu <https://github.com/demotu>, is there a reason you haven't
> checked the performance checkbox in the list in the first comment here?
>
