+++ setup.py
@@ -1,7 +1,6 @@
 from setuptools import setup, find_packages
 from codecs import open
 from os import path
-import re
 
 here = path.abspath(path.dirname(__file__))
 
