### 2017-09-15 f75c766
final writing read through
