### 2017-09-15 #396 @sixpearls
Thanks for your review and feedback, @demotu.

@cMadan, what are the next steps to get this published? And I'm glad you liked the documentation and paper!
