### 2017-09-17 1737286
a packaging/docs adjustment
