+++ docs/overview.rst
@@ -1,3 +1,4 @@
 .. include:: ../README.rst
 
 .. |API documentation| replace:: :doc:`API Documentation<api/api>`
+.. |simupy_logo| replace:: SimuPy
