+++ simupy/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2.0'
+__version__ = '1.0.0'
