### 2017-09-19 #396 @sixpearls
Should the archive of the project include the paper documents? Or just the software package, such as released to PyPi?

Thanks,
Ben
