### 2017-09-19 #396 @sixpearls
This DOI is for the files associated with the software package as distributed by PyPi with the paper subfolder from the repository included additionally. Please let me know if this needs any corrections.

http://doi.org/10.5281/zenodo.896706
